+++ bucket_location.go
@@ -8,15 +8,18 @@ import (
 
 // BucketGetLocationResult ...
 type BucketGetLocationResult struct {
-	XMLName  xml.Name `xml:"LocationConstraint"`
-	Location string   `xml:",chardata"`
+	XMLName xml.Name `xml:"LocationConstraint"`
+	// 说明 Bucket 所在地域，枚举值参见 可用地域[1] 文档，如：ap-beijing、ap-hongkong、eu-frankfurt 等
+	// [1]: https://cloud.tencent.com/document/product/436/6224
+	Location string `xml:",chardata"`
 }
 
 // GetLocation ...
 //
-// Get Bucket Location接口获取Bucket所在地域信息，只有Bucket所有者有权限读取信息。
+// Get Bucket Location 接口用于获取 Bucket 所在的地域信息，该 GET 操作使用 location 参数返回 Bucket 所在的区域，
+// 只有 Bucket 持有者才有该 API 接口的操作权限。
 //
-// https://www.qcloud.com/document/product/436/8275
+// https://cloud.tencent.com/document/product/436/8275
 func (s *BucketService) GetLocation(ctx context.Context) (*BucketGetLocationResult, *Response, error) {
 	var res BucketGetLocationResult
 	sendOpt := sendOptions{