+++ object_test.go
@@ -251,7 +251,7 @@ func TestObjectService_Options(t *testing.T) {
 	})
 
 	opt := &ObjectOptionsOptions{
-		Origin:                     "www.qq.com",
+		Origin: "www.qq.com",
 		AccessControlRequestMethod: "PUT",
 	}